+++ babel.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  presets: [
+    [
+      '@babel/preset-typescript',
+      {
+        targets: {
+          node: '12.16.1',
+        },
+      },
+    ],
+  ],
+};